+++ src/lib.rs
@@ -67,7 +67,7 @@ let sql = "select ename, sal, comm from emp where deptno = :1";
 // Select a table with a bind variable.
 println!("---------------|---------------|---------------|");
 let rows = conn.query(sql, &[&30])?;
-for row_result in &rows {
+for row_result in rows {
     let row = row_result?;
     // get a column value by position (0-based)
     let ename: String = row.get(0)?;
@@ -88,7 +88,7 @@ for row_result in &rows {
 // The rows iterator returns Result<(String, i32, Option<i32>)>.
 println!("---------------|---------------|---------------|");
 let rows = conn.query_as::<(String, i32, Option<i32>)>(sql, &[&10])?;
-for row_result in &rows {
+for row_result in rows {
     let (ename, sal, comm) = row_result?;
     println!(" {:14}| {:>10}    | {:>10}    |",
              ename,